refactor(memory): simplify CO-RE implementation with clean naming
- Remove all legacy code, CO-RE only
- Simplify naming: memory.c instead of memory_monitor_core_simple.c
- Single generate.go file (no separate core/legacy)
- Rename collector_ebpf_core.go to observer_ebpf.go
- Clean struct names: memoryEvent instead of memoryEventCore
- Remove '_core' suffix from all functions
- Generated files now use MemoryObjects pattern

✅ Memory Observer compiles successfully with new structure
✅ Ready to apply same pattern to all observers
