+++ internal/observers/memory/bpf/generate.go
@@ -1,3 +1,4 @@
 package bpf
 
-//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target amd64,arm64 -cc clang memorymonitor ../bpf_src/memory_monitor.c -- -I../../bpf_common -g -O2 -Wall -Wextra
+// CO-RE eBPF generation for Memory Observer
+//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target amd64,arm64 -cc clang Memory ../bpf_src/memory.c -- -I../../bpf_common -g -O2 -Wall -Wextra -Wno-compare-distinct-pointer-types
