+++ internal/observers/memory/observer_ebpf_test.go
